requirements.txt
argcomplete==2.0.0
asgiref==3.6.0
backports.zoneinfo==0.2.1
blinker==1.5
boto3==1.26.70
botocore==1.29.70
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
Django==4.1.6
djangorestframework==3.14.0
durationpy==0.5
dynamorm==0.11.0
hjson==3.1.0
idna==3.4
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.2
marshmallow==3.19.0
packaging==23.0
pip==21.3.1
placebo==0.9.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-slugify==8.0.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
setuptools==60.2.0
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
toml==0.10.2
tqdm==4.64.1
troposphere==4.3.0
tzdata==2022.7
urllib3==1.26.14
Werkzeug==2.2.3
wheel==0.37.1
zappa==0.56.1
